Add uppy backend plugin
